ldso: add null-pointer check
There is a check for (*rpnt == NULL) a few lines above but the "else"
case performing an allocation does only exist if SHARED is not defined.
If SHARED is defined, the allocation is not performed and it may happen
(at least in theory) that *rpnt == NULL when executing

  (*rpnt)->dyn = tpnt;

Add the null-pointer check.

Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>