Makerules: fix out-of-tree build error "Argument list too long"
Make rule that announces and makes static libc archive can generate
command line that is too long when a full path O= option is used. That
breaks the build with the following message:

  make: /bin/sh: Argument list too long

Use make $(file) construct to generate file list file and use that file
as $(AR) argument instead.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>