arm: Replace deprecated asm instructions for ARMv8 AArch32
ARMv8 has particular restrictions which coprocessor can be used and as
such these instructions, which were likely used for backwards
compatibility purposes, cannot be used on ARMv8. We solve this by
checking for ARMv8 and then using the corresponding mnemonics which were
placed in comments alongside the instructions causing issues.

Fixes the following errors:

.../setjmp.S:59:6: error: invalid operand for instruction
 stc p11, cr8, [r12], #68
     ^
.../setjmp.S:62:6: error: invalid operand for instruction
 mrc p10, 7, r2, cr1, cr0, 0
     ^
.../__longjmp.S:69:6: error: invalid operand for instruction
 ldc p11, cr8, [r12], #68
     ^
.../__longjmp.S:73:6: error: invalid operand for instruction
 mcr p10, 7, r1, cr1, cr0, 0
     ^

Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>