bits/stat.h: Declare st_mtim if defined(__USE_XOPEN2K8), everywhere
While building software that sets _POSIX_C_SOURCE=200809L and uses
stat.st_mtim for ARM, it was noticed that st_mtim was not defined.

This seems to be because common/bits/stat.h was picked up, which does
not take __USE_XOPEN2K8 as a reason to enable st_mtim and related
fields. This appears to be an oversight, and porting the check from
common-generic/bits/stat.h to other architectures does indeed fix the
build issue.

This patch is based on commit 50bd6d06e ("Fix memory corruption due to
struct stat field").

Signed-off-by: J. Neuschäfer <j.neuschaefer@gmx.net>