ldso.c: use 'unsigned int' as loop variable type
As the condition for terminating the loop compares the loop variable
against an ElfW() variable which is, depending on the platform, either
uint32_t or uint64_t, use 'unsigned int' rather than 'int' for the loop
variable to prevent corresponding compiler warnings.

Note that it would not make sense to use 'unsigned long' because the
number of program headers will never exceed 32-bit.

Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>